+++ packages/jss/src/plugins/fontFaceRule.js
@@ -2,8 +2,10 @@
 import toCss from '../utils/toCss'
 import type {CSSFontFaceRule, RuleOptions, JssStyle, ToCssOptions, BaseRule} from '../types'
 
-export default class FontFaceRule implements BaseRule {
+export class FontFaceRule implements BaseRule {
   type = 'font-face'
+
+  at: string = '@font-face'
 
   key: string
 
@@ -35,5 +37,11 @@ export default class FontFaceRule implements BaseRule {
     }
 
     return toCss(this.key, this.style, options)
+  }
+}
+
+export default {
+  onCreateRule(key: string, style: JssStyle, options: RuleOptions): FontFaceRule | null {
+    return key === '@font-face' ? new FontFaceRule(key, style, options) : null
   }
 }