+++ packages/jss/src/renderers/DomRenderer.js
@@ -6,6 +6,7 @@ import type {
   CSSStyleRule,
   CSSMediaRule,
   CSSKeyframesRule,
+  CSSKeyframeRule,
   Rule,
   ContainerRule,
   JssValue,
@@ -32,7 +33,10 @@ const memoize = <Value>(fn: () => Value): (() => Value) => {
 /**
  * Get a style property value.
  */
-function getPropertyValue(cssRule: HTMLElement | CSSStyleRule, prop: string): string {
+function getPropertyValue(
+  cssRule: HTMLElement | CSSStyleRule | CSSKeyframeRule,
+  prop: string
+): string {
   try {
     // Support CSSTOM.
     if ('attributeStyleMap' in cssRule) {
@@ -49,7 +53,11 @@ function getPropertyValue(cssRule: HTMLElement | CSSStyleRule, prop: string): st
 /**
  * Set a style property.
  */
-function setProperty(cssRule: HTMLElement | CSSStyleRule, prop: string, value: JssValue): boolean {
+function setProperty(
+  cssRule: HTMLElement | CSSStyleRule | CSSKeyframeRule,
+  prop: string,
+  value: JssValue
+): boolean {
   try {
     let cssValue = ((value: any): string)
 
@@ -79,7 +87,7 @@ function setProperty(cssRule: HTMLElement | CSSStyleRule, prop: string, value: J
 /**
  * Remove a style property.
  */
-function removeProperty(cssRule: HTMLElement | CSSStyleRule, prop: string) {
+function removeProperty(cssRule: HTMLElement | CSSStyleRule | CSSKeyframeRule, prop: string) {
   try {
     // Support CSSTOM.
     if ('attributeStyleMap' in cssRule) {
@@ -347,8 +355,9 @@ export default class DomRenderer {
     const {sheet} = this.element
 
     if (rule.type === 'conditional' || rule.type === 'keyframes') {
-      const containerRule = ((rule: any): ContainerRule)
-      const cssRule = insertRule(sheet, `${containerRule.key} {}`, index)
+      const containerRule: ContainerRule = (rule: any)
+      // We need to render the container without children first.
+      const cssRule = insertRule(sheet, containerRule.toString({children: false}), index)
       if (cssRule === false) {
         return false
       }
@@ -356,7 +365,6 @@ export default class DomRenderer {
         const cssChildRule = insertRule(cssRule, childRule.toString(), childIndex)
         if (cssChildRule !== false) childRule.renderable = cssChildRule
       })
-
       return cssRule
     }